src/requirements-lambda.txt
# Lightweight dependencies for the API Lambda handler.
# boto3 is already provided by the Lambda runtime.
loguru>=0.7.3
pymongo>=4.15.2
certifi>=2024.0.0
lxml>=6.0.2
Pillow>=10.4.0
